# activity/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "OpenShock.Activity.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenShockProblem"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenShockProblem"
                }
              }
            }
          }
        }
      }
    },
    "/dev/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevTokenRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DevTokenRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DevTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevTokenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthUserDto": {
        "required": [
          "discordId",
          "name",
          "avatar"
        ],
        "type": "object",
        "properties": {
          "discordId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatar": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "DevTokenRequest": {
        "required": [
          "discordId",
          "name"
        ],
        "type": "object",
        "properties": {
          "discordId": {
            "type": "string"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "DevTokenResponse": {
        "required": [
          "jwt"
        ],
        "type": "object",
        "properties": {
          "jwt": {
            "type": "string"
          }
        }
      },
...
      "OpenShockProblem": {
        "required": [
          "type",
          "title"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "requestId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TokenRequest": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
      "TokenResponse": {
        "required": [
          "discordAccessToken",
          "jwt",
          "user"
        ],
        "type": "object",
        "properties": {
          "discordAccessToken": {
            "type": "string"
          },
          "jwt": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        }
      },
...
    }
  },
...
}